Second major refactor of streams.
This one was a class by class examination of the transformation package.  A couple things moved to other packages, a couple things were deleted, and patterns were modified to get rid of streams references.  Now you should be able to run the transformations completely independently of the tracing/control structures in streams.