Rewrite java runtime to generate transformations with a parser instead of the compiler.
This isn't pretty, but I need to check something in since it's too large now.  There are a lot of refactors and pushes of stuff around here, but the primary change here is to change things so that transformations don't need to compile to be turned into descriptions of the transformation.  This allows us to not have dependency versions resolved.  That way versions can be either determined upon load (for bootstrapping) or plugged into the current transformation dependencies when the code is updated.

It's a bit of controversial change since it makes things more complex even after I get things cleaned up.  Parse trees are not very fun.